change some type to remove the warning.

+++ trunk/libproject/Utility/strutility.cpp
@@ -6,27 +6,27 @@
 namespace strutility {
 
 void trim_right(TCHAR * str, const TCHAR* trimed) {
-	const int len = _tcslen(str);
-	for (int i = len-1; i >=0 && _tcschr(trimed, str[i]); ++i) {
+	const size_t len = _tcslen(str);
+	for (size_t i = len-1; i >=0 && _tcschr(trimed, str[i]); ++i) {
 		str[i] = '\0';
 	}
 }
 
 void trim_left(TCHAR * str, const TCHAR* trimed) {
-	const int len = _tcslen(str);
-	int need_to_move;
+	const size_t len = _tcslen(str);
+	size_t need_to_move;
 	for (need_to_move = 0; need_to_move < len &&
 			_tcschr(trimed, str[need_to_move]); ++need_to_move) {
 		;
 	}
 
 	// 先前移动
-	for (int i = 0; i < len - need_to_move; i++) {
+	for (size_t i = 0; i < len - need_to_move; i++) {
 		str[i] = str[i + need_to_move] ;
 	}
 	
 	// 剩余的空间至0
-	for (int i = 0; i < need_to_move; i++) {
+	for (size_t i = 0; i < need_to_move; i++) {
 		str[len-i-1] = '\0';
 	}
 }
